requirements.txt
aiogram==3.*
python-dotenv
python-telegram-bot